+++ effect.cpp
@@ -796,7 +796,7 @@ int32 effect::in_range(const chain& ch) {
 void effect::set_activate_location() {
 	card* phandler = get_handler();
 	active_location = phandler->current.location;
-	active_sequence = phandler->current.sequence;
+	active_sequence = static_cast<uint16>(phandler->current.sequence);
 }
 void effect::set_active_type() {
 	card* phandler = get_handler();
